Add test custom template for test
This template is used to link the static js code bundles, as well
as providing a div for the frontend framework to render in.

This way we can use a normal Django view that serves this template
and allows a SPA to be rendered within a Django template.
Separately in the JS code you want to mount the app using the
div in the template