Edit Album Create API view to handle front end data sent from client
Added MultiPartParser to parse mutlipart/form-data

Edited post method for API view to handle form with file and json data.
The serializer expects a dictionary object for certain fields so
json.loads is used to convert the text/json into a dict for those fields.

The files also were sent as a separate member in the form data, and is attached
to the appropriate dictionary object before serialization.