Initial Album Creation Page for user
This page provides the user with a form to upload a new album.
The initial functionality is finished and the user can upload an almbum from this page.

TODO: Add CSS to page. Add ability for user to create new Labels/Artists/Genres in a simple
way so that they can provide these new values for the Album they are uploading.