Reformatted AlbumCreationPageJSX and some functions. Added tailwind css for basic initial formatting
JSX and tailwind classes were added to format the form.

Checkbox component was utilized instead of select html element.
This allows for a cleaner way for the user to choose artists/genres.

TODO:

Continue styling with css.

Add useEffect and useState hooks to fetch Artists/Genres/Labels to display
for selection by the user instead of dummy data.

Add button to add new Artist/Genre/Label for the user, and then fetch again
and re-render so new Artist/Genre/Label is present for selection by user.